add `writev` method to FileHandle class
